Separate test helpers.
